+++ crates/extrinsics/src/instantiate.rs
@@ -34,9 +34,9 @@ use super::{
     PairSigner,
     StorageDeposit,
     TokenMetadata,
+    DEFAULT_KEY_COL_WIDTH,
     MAX_KEY_COL_WIDTH,
 };
-use crate::DEFAULT_KEY_COL_WIDTH;
 use anyhow::{
     anyhow,
     Context,
